Refactor Atom Map Mixin
